Move ProfileArgs to schemas
Generated-by: claude-code Opus 4.6
